// apps/web/lib/api/generated/models/chatPinnedItemResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * llame API
 * llame auth and domain API
 * OpenAPI spec version: 0.1
 */
import type { ChatPinnedItemResponseItemType } from "./chatPinnedItemResponseItemType";
import type { ChatRefCard } from "./chatRefCard";

export interface ChatPinnedItemResponse {
  itemType: ChatPinnedItemResponseItemType;
  itemId: string;
  pinnedAt: string;
  item: ChatRefCard;
}