// apps/web/lib/api/generated/models/orgUnitConflictErrorResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * llame API
 * llame auth and domain API
 * OpenAPI spec version: 0.1
 */
import type { OrgUnitConflictErrorResponseCode } from "./orgUnitConflictErrorResponseCode";

export interface OrgUnitConflictErrorResponse {
  statusCode: number;
  error: string;
  message: string;
  code: OrgUnitConflictErrorResponseCode;
}