// apps/web/lib/api/generated/models/projectPinnedItemResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * llame API
 * llame auth and domain API
 * OpenAPI spec version: 0.1
 */
import type { ProjectPinnedItemResponseItemType } from "./projectPinnedItemResponseItemType";
import type { ProjectRefCard } from "./projectRefCard";

export interface ProjectPinnedItemResponse {
  itemType: ProjectPinnedItemResponseItemType;
  itemId: string;
  pinnedAt: string;
  item: ProjectRefCard;
}